Implement own navigation logic
