+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/compose/LoginView.kt
@@ -10,25 +10,16 @@ import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
-import de.hsfl.budgetBinder.domain.use_case.auth_user.LoginUseCase
 import de.hsfl.budgetBinder.presentation.LoginState
-import de.hsfl.budgetBinder.presentation.LoginViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.SupervisorJob
-import org.kodein.di.compose.localDI
-import org.kodein.di.instance
 
 @Composable
-fun LoginView() {
-    val scope = CoroutineScope(Dispatchers.Unconfined + SupervisorJob())
-    val di = localDI()
-    val loginUseCase: LoginUseCase by di.instance()
-    val viewModel = LoginViewModel(loginUseCase, scope)
-    val state by viewModel.state.collectAsState(scope)
-
+fun LoginView(
+    onLoginButtonPressed: (email: String, password: String) -> Unit,
+    state: State<Any>
+) {
     var emailTextFieldState by remember { mutableStateOf("") }
     var passwordTextFieldState by remember { mutableStateOf("") }
+    val viewState by remember { state }
 
     MaterialTheme {
         Column(
@@ -52,24 +43,26 @@ fun LoginView() {
                 keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password)
             )
 
-            when(state) {
+            /*when (viewState) {
                 is LoginState.Success -> {
-                    Text((state as LoginState.Success).login.toString())
+                    Text((viewState as LoginState.Success).login.toString())
                 }
                 is LoginState.Error -> {
-                    Text(text = (state as LoginState.Error).error,
-                    color = MaterialTheme.colors.error,
-                    fontWeight = FontWeight.Bold
+                    Text(
+                        text = (viewState as LoginState.Error).error,
+                        color = MaterialTheme.colors.error,
+                        fontWeight = FontWeight.Bold
                     )
                 }
                 is LoginState.Loading -> {
                     CircularProgressIndicator()
                 }
-            }
+            }*/
 
             Button(onClick = {
                 localFocusManager.clearFocus()
-                viewModel.auth(emailTextFieldState, passwordTextFieldState)
+                onLoginButtonPressed(emailTextFieldState, passwordTextFieldState)
+                //viewModel.auth(emailTextFieldState, passwordTextFieldState)
 
             }) {
                 Text("Login")
