+++ budget-binder-multiplatform-app/src/jvmMain/kotlin/de/hsfl/budgetBinder/compose/Screen.kt
@@ -1,0 +1,6 @@
+package de.hsfl.budgetBinder.compose
+
+sealed class Screen {
+    object Login : Screen()
+    object User : Screen()
+}
